orc-requirements.txt
Faker==19.6.1
fastapi==0.103.1
httpx==0.25.0
pydantic-settings==2.0.3
python-dotenv==1.0.0
redis==5.0.0
starlette==0.27.0
uvicorn==0.23.2